+++ plugin_qgis_lpo/action_scripts/joke.py
