+++ plugin_qgis_lpo/processing/message.html
@@ -1,0 +1,38 @@
+<font style="font-size:18px"><b>Besoin d'aide ?</b> Vous pouvez vous référer au <b>Wiki</b> accessible sur ce lien : <a
+        href="https://github.com/lpoaura/PluginQGis-LPOData/wiki"
+        target="_blank">https://github.com/lpoaura/PluginQGis-LPOData/wiki</a>.</font><br /><br />
+Cet algorithme vous permet, à partir des données d'observation enregistrées dans la base de données LPO, d'obtenir un
+<b>tableau de synthèse</b> par espèce (couche PostgreSQL) basé sur une <b>zone d'étude</b> présente dans votre projet
+QGis (couche de type polygones).
+<b style='color:#952132'>Les données d'absence sont exclues de ce traitement.</b><br /><br />
+<b>Pour chaque espèce</b> observée dans la zone d'étude considérée, le tableau fournit les informations suivantes :
+<ul>
+    <li>Identifiant VisioNature de l'espèce</li>
+    <li>cd_nom et cd_ref</li>
+    <li>Rang</li>
+    <li>Groupe taxonomique auquel elle appartient</li>
+    <li>Nom français de l'espèce</li>
+    <li>Nom scientifique de l'espèce</li>
+    <li>Nombre de données</li>
+    <li>Nombre de données / Nombre de données TOTAL</li>
+    <li>Nombre d'observateurs</li>
+    <li>Nombre de dates</li>
+    <li>Nombre de données de mortalité</li>
+    <li>LR France</li>
+    <li>LR régionale(s)</li>
+    <li>Directive Habitats</li>
+    <li>Directive Oiseaux</li>
+    <li>Protection nationale</li>
+    <li>Statut nicheur (pour les oiseaux)</li>
+    <li>Nombre d'individus maximum recensé pour une observation</li>
+    <li>Année de la première observation</li>
+    <li>Année de la dernière observation</li>
+    <li>Liste des communes</li>
+    <li>Liste des sources VisioNature</li>
+</ul>
+<br />
+<font style='color:#0a84db'><u>IMPORTANT</u> : Les <b>étapes indispensables</b> sont marquées d'une <b>étoile *</b>
+    avant leur numéro. Prenez le temps de lire <u>attentivement</U> les instructions pour chaque étape, et
+    particulièrement les</font>
+<font style='color:#952132'>informations en rouge</font>
+<font style='color:#0a84db'>!</font>
